refactor omemo fingerprint UI code
