+++ art/render.rb
@@ -50,6 +50,7 @@ images = {
 	'ic_notifications_off_white80.svg' => ['ic_notifications_off_white80', 24],
 	'ic_notifications_paused_white80.svg' => ['ic_notifications_paused_white80', 24],
 	'ic_notifications_white80.svg' => ['ic_notifications_white80', 24],
+	'ic_verified_fingerprint.svg' => ['ic_verified_fingerprint', 36],
 	'md_switch_thumb_disable.svg' => ['switch_thumb_disable', 48],
 	'md_switch_thumb_off_normal.svg' => ['switch_thumb_off_normal', 48],
 	'md_switch_thumb_off_pressed.svg' => ['switch_thumb_off_pressed', 48],
@@ -57,7 +58,7 @@ images = {
 	'md_switch_thumb_on_pressed.svg' => ['switch_thumb_on_pressed', 48],
 	'message_bubble_received.svg' => ['message_bubble_received.9', 0],
 	'message_bubble_received_grey.svg' => ['message_bubble_received_grey.9', 0],
-   'message_bubble_received_dark.svg' => ['message_bubble_received_dark.9', 0],
+	'message_bubble_received_dark.svg' => ['message_bubble_received_dark.9', 0],
 	'message_bubble_received_warning.svg' => ['message_bubble_received_warning.9', 0],
 	'message_bubble_received_white.svg' => ['message_bubble_received_white.9', 0],
 	'message_bubble_sent.svg' => ['message_bubble_sent.9', 0],
