+++ src/main/res/menu/omemo_key_context.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/purge_omemo_key"
+        android:title="@string/purge_key"/>
+    <item
+        android:id="@+id/copy_omemo_key"
+        android:title="@string/copy_fingerprint"/>
+</menu>
