+++ src/main/res/values/ids.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item type="id" name="TAG_ACCOUNT"/>
+    <item type="id" name="TAG_FINGERPRINT"/>
+</resources>
